refactor directory
