+++ bifrost/bifrost.go
@@ -1,8 +1,8 @@
 package bifrost
 
 import (
-	"github.com/Mintegral-official/mtggokit/data/bifrost/streamer"
-	"github.com/Mintegral-official/mtggokit/data/container"
+	"github.com/Mintegral-official/mtggokit/bifrost/container"
+	"github.com/Mintegral-official/mtggokit/bifrost/streamer"
 	"github.com/pkg/errors"
 )
 
@@ -12,19 +12,19 @@ type Logger interface {
 	Warnf(format string, v ...interface{})
 }
 
-type Loader struct {
+type Bifrost struct {
 	DataStreamers map[string]streamer.DataStreamer
 	sched         Sched
 	logger        *Logger
 }
 
-func NewLoader() *Loader {
-	return &Loader{
+func NewLoader() *Bifrost {
+	return &Bifrost{
 		DataStreamers: make(map[string]streamer.DataStreamer),
 	}
 }
 
-func (l *Loader) Get(name string, key container.MapKey) (interface{}, error) {
+func (l *Bifrost) Get(name string, key container.MapKey) (interface{}, error) {
 	s, ok := l.DataStreamers[name]
 	if !ok {
 		return nil, errors.New("not found streamer[" + name + "]")
@@ -36,16 +36,15 @@ func (l *Loader) Get(name string, key container.MapKey) (interface{}, error) {
 	return c.Get(key)
 }
 
-func (l *Loader) Register(name string, streamer streamer.DataStreamer) error {
+func (l *Bifrost) Register(name string, streamer streamer.DataStreamer) error {
 	if _, ok := l.DataStreamers[name]; ok {
 		return errors.New("streamer[" + name + "] has already exist")
 	}
 	l.DataStreamers[name] = streamer
-	l.sched.AddStreamer(name, streamer)
 	return nil
 }
 
-func (l *Loader) GetStreamer(name string) (streamer.DataStreamer, error) {
+func (l *Bifrost) GetStreamer(name string) (streamer.DataStreamer, error) {
 	s, ok := l.DataStreamers[name]
 	if !ok {
 		return nil, errors.New("not found streamer[" + name + "]")
