+++ bifrost/container/buffered_map_container.go
