+++ bifrost/schedule_test.go
@@ -2,7 +2,7 @@ package bifrost
 
 import (
 	"context"
-	"github.com/Mintegral-official/mtggokit/data/bifrost/streamer"
+	"github.com/Mintegral-official/mtggokit/bifrost/streamer"
 	"github.com/smartystreets/goconvey/convey"
 	"testing"
 	"time"
